+++ packages/aws-durable-execution-sdk-js-insight-desktop/.gitignore
@@ -1,0 +1,3 @@
+dist/
+release/
+node_modules/
