+++ packages/aws-durable-execution-sdk-js-insight-desktop/tsconfig.test.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "types": ["node", "jest"]
+  },
+  "include": ["src"],
+  "exclude": []
+}
